otex-core:   source/java/com/google/enterprise/connector/otex/ 	LivelinkQueryTraversalManager.java 	LivelinkConnector.java 	LivelinkResultSet.java   config/ 	connectorInstance.xml
  Added a property "startDate" to connectorInstance.xml which
indicates the date at which to start traversals.
  LivelinkConnector.setStartDate() tries to parse this date as both
"default" and "SHORT" date types.  If this fails, a warning is issued,
but searching proceeds.  (Is this proper behavior?)  An empty string
starts at the beginning of time with no warning.

otex-lapi:
	source/javatests/com/google/enterprise/connector/otex/LivelinkTest.java

  Added a test that sets a start date and verifies both that there is
at least one document older than that in the repository and that none
older are returned in a traversal.