Fixed issue 18. When there are starting nodes, getExcluded in LivelinkTraversalManager now only looks up the tree to just below the starting nodes, which themselves might be hidden.
Fixed an error in the core SQL predicate, which was qualifying the
candidates predicate with "T", but that's the DTree fields for the
ancestors in this case. Instead, we need to be qualifying it with
"Anc", which includes the child rows as the candidates, and makes the
ancestor rows as the ones were are checking for hidden items.