Fixed delete checkpoints against SQL Server.
DAuditNew.EventID is a NUMBER(10) in Oracle, and returned by ListNodes
as an LLValue.LL_INTEGER. It is a NUMERIC(10,0) in SQL Server, and
returned by ListNodes as an LLValue.LL_DOUBLE. Neither return type
seems particularly correct (for example, an integer can't hold all the
values of a NUMBER(10)). Casting the SQL Server EventID column to a
BIGINT appears to return a long (the integral values are prefixed with
"L" in the ListNodes output), but LAPI doesn't support longs.
Strangely, the error appears trying unmarshall the status, which is
fine, rather than the EventID values.

Major changes:

* Removed the CASTs from the SQL queries, and let ListNodes return
  doubles for SQL Server.
  
* Instead of passing the EventID to Checkpoint.setDeleteCheckpoint as
  a string and calling Long.parseLong on it, we now pass the EventID
  as a ClientValue, and either call toInteger on it, or call toDouble
  on it and cast the result to a long, depending on the type.

Other changes:

* Deleted the unused overloads setInsertCheckpoint(Date, String) and
  setDeleteCheckpoint(Date, long).

* Formatted LivelinkTraversalManager.getDeletesOracle to match the
  formatting changes to getDeletesSqlServer in r267.