The runtime check was looking at a constant from SpiConstants that was compiled into the connector, so it would fail to compile, but it would not fail at runtime. Switched to explicitly using reflection to check for the field.
When the connector relies on a class from the CM, as it does now with
UrlValidator, this test has no effect. When it relies on a constant,
typically one of the property name constants, or on some invisible
connector manager behavior, then the check is more useful.

http://codereview.appspot.com/4176050/

M    projects/otex-core/source/java/com/google/enterprise/connector/otex/LivelinkConnector.java