Fix issue 7005056: autoDetectServtype query fails against OTCS 10.
This is a duplicate of issue 6032685, but fixing this has minor merit
of its own. It's the last query without a test, and the fix here
removes the need to whitelist the KDual and dual tables.

The problem with OTCS 10 is that the column names are quoted. In this
case, the only select column was 42, which is invalid quoted. The fix
here selects DataID from DTree (with no rows, for efficiency).

Code review: https://codereview.appspot.com/8664044/