Fix issue 6513973, part one: Unify SQL handling in LivelinkTraversalManager.
Each query is in three parts, as required by the LAPI ListNodes method.
The resource bundle key is of the form: "className.methodName.part",
where part is one of "select", "from", or "where".

I may want to keep the Oracle and SQL Server variations closer together
in the code, so they are easier to compare, but I'm not doing that yet.

The executeLimit method is designed to work with both Oracle's ROWNUM
and SQL Server's TOP. It may go away in a subsequent CL, because OTCS
10 does not support the use of TOP in the outer select list.

Not everything has moved to the resource bundles. The choice of the
select list and views for the results query, and the full select list,
are still part of the LivelinkTraversalManager class.

Code review: https://codereview.appspot.com/8879043/