Fixed issue 9923703: Livelink connector may skip documents when documents are updated during a batch and transient errors occur.
The results queries, getMatching and getMatchingDescendants, now filter
out documents where the ModifyDate is later than the last ModifyDate
in the candidates query results. This maintains the strict date order
of processing, and delays processing newly updated documents until the
checkpoint gets to the current date.

All calls to getMatching filter out later results, even when a
genealogist is used and getMatchingDescendants will be called later.
This is just to avoid the extra work of the genealogist on results we
will filter later, although newly updated documents are most likely to
happen with getMatchingDescendants, since the slower genealogists will
delay that query.

I added a new test document, object ID 66, which is "updated" by
simple virtue of having a later ModifyDate in the WebNodes table used
by the final results query than in the DTree table used by the
candidates query.

Other changes:

* Refactored 15 calls to getMatchingQuery(null, new Date(), false)
  into a private helper method.

* Added checks for non-null DocumentLists. This didn't used to work,
  but now does, apparently (probably something in the query
  refactoring for OTCS 10).

Because the queries have changed, we should reverify the execution
plans in Oracle and SQL Server. I don't have access at the moment, but
I am committing this since it fixes a correctness issue. We should
check the plans before including this change in a general release.

Code review: https://codereview.appspot.com/11525048