Add a retry for a bad candidates set. At a customer site, under load, apparently Oracle is returning the wrong result set once every few thousand executions of the candidates query. The retry mechanism is simply to throw a RepositoryException, so if the bad candidates sets keep coming, we'll continue retrying the query.
The check for a bad result set is disabled by default, and configured
with the candidatesTimeWarpFuzz advanced configuration property. A
negative value (-1 by default) disables the check. Zero checks only
for results older than the checkpoint, or the same age but with a
smaller DataID, which is always wrong. A positive value, N > 0, checks
for results more than N days newer than the checkpoint, which is fine
for small values of N, but unlikely to be right for larger values.

Code review: https://codereview.appspot.com/13242045/