Replace assertions with Guava Preconditions checks. The assertions are on live data, so we should use checks that cannot be disabled in production. They were also missing diagnostic messages with the failed inputs.
This is related to issue 6888984, where one of the assertions would
have fired but was disabled (and without a message the assertion was
nearly useless in any case).

There are other assertions in the connector, but only on hard-coded
data that is verified by the tests.

Code review: https://codereview.appspot.com/58890043/