Optionally use DTreeAncestors in getCandidates (b/6514008).
For better performance with Items to Index, use DTreeAncestors in the
initial candidates query to avoid scanning everything in DTree.

Add a new advanced config property, useDTreeAncestorsFirst. It defaults
to false, which preserves the existing behavior of using DTreeAncestors
in the results query, rather than the initial candidates query. If the
new property is set to true, useDTreeAncestors must also be set to true
(which it is by default).

use..First = false, use... = true => old behavior (default)
use..First = false, use... = false => genealogist
use..First = true, use... = true => new behavior
use...First = true, use... = false => configuration error

With the new behavior, the core candidates query changes from

    select ModifyDate,DataID
    from DTree
    [where ModifyDate > ? or (ModifyDate = ? and DataID > ?)]
    order by ModifyDate, DataID
to
    select ModifyDate,DataID
    from DTree [join DTreeAncestors using (DataID)
    where (AncestorID in (?) or DataID in (?))]
    [{and | where} (ModifyDate > ? or (ModifyDate = ? and DataID > ?))]
    order by ModifyDate, DataID

For example, if the Items to Index contains "2000", then the WHERE
clause restriction is

    AncestorID in (2000,-2000) or DataID in (2000)

This matches the query in getDescendants. This is needed for Livelink 9,
but not for OTCS 10. The issue is that in Livelink 9, DTreeAncestors
does not include an <objectID, objectID> row, for example, <2000, 2000>,
and it does not include the stepparents of volume nodes, for example,
Projects. So the -2000 checks for the volume node along with the start
node, and we explicitly check DataID for the start nodes.

In OTCS 10, both the stepparent and self entries are included in
DTreeAncestors (though the -1 entries in Livelink 9 are not).

Note: SQL Server does not support USING, but I eschew consistency since
for Oracle the join is part of the whitelisted view, and I want that to
be as simple as possible.

Changes in the tests:

* For reasons lost to time, entries were missing from WebNodes (for
  object IDs 6 and 2000) and DTreeAncestors (children of 6 were missing
  entries with AncestorID = 2000). Adding these broke a few tests in
  trivial ways. DTreeAncestors in the tests now contains the
  intersection of rows from Livelink 9 and OTCS 10, eliding both the
  self and -1 entries.

* Improved tests by comparing results using expected objectIDs, rather
  than just the expected number of results.

* H2 also does not support USING, so more query rewriting was needed in
  MockClient.

The unit tests pass with both Oracle and SQL Server queries.

Code review: https://codereview.appspot.com/239230043/