Wrap LivelinkTraversalManager with a class than manages the clients.
Workaround a bug in LAPI's HTTP tunneling (1 of 2; b/9021325).
Part 1 is for the TraversalManager, part 2 will be for the Lister.

The error handling in LAPI is broken when using HTTP tunneling. The
fix is to get a new LLSession when using the previous one throws an
exception. Detecting those exceptions would likely be very invasive,
and creating a new LLSession is not super heavyweight (though it is
likely more work with HTTP tunneling), so I just get a new LLSession
for each batch. Getting a new LLSession requires getting a new Client,
because all of the other LAPI objects are dependent on the LLSession.

The clients are injected into LivelinkTraversalManager for
testability, so changing the clients inside the class would break
dependency injection. To preserve that, I'm wrapping the existing
LivelinkTraversalManager class in another TraversalManager
implementation that gets a new LivelinkTraversalManager when it wants
a new client.

I could inject a factory that had a strategy for when to actually
create new clients, but I think this hard-coded wrapper is much
simpler, and it changes less existing code. It also avoids introducing
non-final fields in the core of the connector. I am planning to inject
a factory to fix the issue in the Lister.

Related change:

* Move the getTraversalManager, getCurrentUsername, and impersonateUser
  methods from LivelinkSession to TraversalManagerWrapper unchanged.

Tested manually to see that the new code recovers from previously
permanent errors when Livelink is restarted or the CGI crashes.

Code review: https://codereview.appspot.com/260990043/