Inject a ClientFactory instead of a Client into GroupAdaptor.
Workaround a bug in LAPI's HTTP tunneling (2 of 2; b/9021325).
Part 1 was for the TraversalManager, part 2 is for the Lister.

The error handling in LAPI is broken when using HTTP tunneling. The
fix is to get a new LLSession when using the previous one throws an
exception. Detecting those exceptions would likely be very invasive,
and creating a new LLSession is not super heavyweight (though it is
likely more work with HTTP tunneling), so I just get a new LLSession
for each listing. Getting a new LLSession requires getting a new
Client, because all of the other LAPI objects are dependent on the
LLSession.

Tested manually to see that the new code recovers from previously
permanent errors when Livelink is restarted or the CGI crashes.

Code review: http://codereview.appspot.com/263100043