Add caching to avoid sending duplicate deletes every second.
This is a fix for commit 7f0689e. The useIndexedDeleteQuery uses
AuditDate >= {0}, so subsequent batches will always return the last
delete entry in the audit trail (and possibly more). That prevents
the DocumentList from ever being null, which prevents the CM from
ever delaying for the retry interval.

This change keeps a cache of the deletes in the previous batch,
checks to see if all the deletes are cached, so that resumeTraversal
can return null, and filters out cached deletes in nextDocument, to
avoid sending duplicate deletes to the GSA.

Code review: http://codereview.appspot.com/313050043/