Avoid resetting the cache when the next cache is empty.
When caching deletes, there should always be at least one cached
delete (if we've ever seen a delete), so if nextDelCache is empty, it
doesn't reflect reality. For example, maybe all deletes were
previously cached so delArray is null, or this batch has timed out and
not processed all the rows.

Other test changes:

* Rename testGetDeletes* to testResumeTraversalDeletes*, since we
  changed the test in commit 05457c1 to call resumeTraversal rather
  than the internal getDeletes method.

* Since getDocids reads the document list, the extra loop to do so was
  pointless (and would get null on the first call to nextDocument).

* Add a count-limited getDocids overload, and assertNullDocumentList.